Decompose danger checks into separate rules
