+++ .flowconfig
@@ -5,6 +5,7 @@
 # Tracking https://github.com/facebook/flow/issues/4015
 <PROJECT_ROOT>/docs
 <PROJECT_ROOT>/test
+<PROJECT_ROOT>/danger
 
 [options]
 suppress_comment= \\(.\\|\n\\)*\\$FlowFixMe
